Moved Buttons: - I realized that I put the buttons inside the LaunchScreen.Storyboard file, so I deleted the buttons from the LaunchScreen file and moved them to the Main.storyboard.   - I edited the buttons in the Main.storyboard to be the right text and changed the color of the buttons and text.
Added Actions to Veiw Controller:
I tried to add all the buttons to the view controller with one IBAction, although I think I only was able to attach the 6 button. This is because the 6 button is the only button that works on the UI.
I also added two lines of code to change the labels text to a String of the buttons being pressed, which has been unwrapped to a Double value.